Reorganising and open PPT
