+++ Global/Utils/GeneratorUtils.cs
@@ -1,0 +1,11 @@
+namespace OpenXMLOffice.Global;
+
+public static class GeneratorUtils
+{
+
+    public static string GenerateNewGUID()
+    {
+        return string.Format("{{{0}}}", Guid.NewGuid().ToString("D").ToUpper());
+    }
+
+}
