+++ Presentation/Internal/CommonSlideData.cs
