+++ Presentation/Internal/Slide.cs
