+++ Presentation/Internal/SlideLayout.cs
