+++ Presentation/Internal/SlideMaster.cs
