+++ Presentation/Internal/ViewProperties.cs
