--- Presentation/PresentationCore.cs
@@ -1,5 +1,0 @@
-namespace OpenXMLOffice.Presentation;
-internal class PresentationCore
-{
-
-}
